UPT: Add tomcat download trap

+++ novnc/config/nobody/install.sh
@@ -70,12 +70,17 @@ sudo -H -u novnc mkdir -p /home/novnc/guacamole
 
 # setup tomcat
 sudo -H -u novnc mkdir -p /home/novnc/tomcat
-sudo -H -u novnc wget https://dlcdn.apache.org/tomcat/tomcat-9/v9.0.64/bin/apache-tomcat-9.0.64.tar.gz
-sudo -H -u novnc tar xvzf apache-tomcat-9.0.64.tar.gz --strip-components 1 --directory /home/novnc/tomcat/
-rm -rf apache-tomcat-9.0.64.tar.gz
-chmod +x /home/novnc/tomcat/bin/*.sh
-mkdir -p /home/novnc/tomcat/webapps
+sudo -H -u novnc wget https://dlcdn.apache.org/tomcat/tomcat-9/v9.0.64/bin/apache-tomcat-9.0.64.tar.gz -O /home/novnc/apache-tomcat-9.0.64.tar.gz
 
+if [ -f /home/novnc/apache-tomcat-9.0.64.tar.gz ]; then
+  sudo -H -u novnc tar xvzf /home/novnc/apache-tomcat-9.0.64.tar.gz --strip-components 1 --directory /home/novnc/tomcat/
+  rm -rf apache-tomcat-9.0.64.tar.gz
+  chmod +x /home/novnc/tomcat/bin/*.sh
+  mkdir -p /home/novnc/tomcat/webapps
+else
+  echo "Failed to download tomcat, please investigate"
+  exit 1;
+fi
 
 # setup guacamole-client
 # --> we will place a predownloaded war from: wget https://dlcdn.apache.org/guacamole/1.4.0/binary/guacamole-1.4.0.war -O /opt/tomcat/webapps/guacamole.war
